pyproject.toml
[project]
name = "proxy-pool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "apscheduler>=3.11.2",
    "fastapi>=0.128.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "uvicorn[standard]>=0.40.0",
]
